requirements.txt
mysql-connector-python==8.3.0
python-dotenv==1.0.0
PyQt6==6.5.3
PyQt6-Qt6==6.5.3
PyQt6-sip==13.6.0
tabulate==0.9.0
colorama==0.4.6 